requirements.txt
# ── Core Django ───────────────────────────────────────────────────────────────
Django>=5.0,<6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# ── Authentication ────────────────────────────────────────────────────────────
djoser==2.3.3
django-cors-headers==4.3.1 
social-auth-app-django==5.7.0
social-auth-core==4.8.5
pyjwt[crypto]==2.11.0
python3-openid==3.2.0
oauthlib==3.3.1
requests-oauthlib==2.0.0
cryptography==46.0.5

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.11

# ── API Utilities ─────────────────────────────────────────────────────────────
python-decouple==3.8
django-filter==24.2
drf-spectacular==0.27.2
Pillow==10.3.0

# ── Cache ─────────────────────────────────────────────────────────────────────
django-redis==5.4.0

# ── Production server ─────────────────────────────────────────────────────────
gunicorn==22.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
requests==2.32.5